Avoid leaving orphaned nodes in the database after testing
Each execution of test 005_db.t creates two unique nodes in the database, only one of which is deleted at the end. Running the test suite multiple times on the same database may result in a large number of nodes with no label, no properties and no relations to other nodes. This may be a source of confusion for developers debugging unrelated database code.

This change attempts to make sure all testing nodes are removed.