Drop dependencies on JSON, JSON::XS, JSON::ize
JSON::XS has been failing to install on blead Perl for half a year (v5.41). It doesn't seem like a good idea to depend upon it.

JSON::MaybeXS is a modern replacement for JSON.pm. It can use JSON::XS and JSON::PP when available, but actually prefers Cpanel::JSON::XS, which is a long-proven and well-maintained fork of JSON::XS.

JSON::MaybeXS was already an indirect prerequisite to REST::Neo4p because Neo4j::Driver depends upon it. JSON::ize did, too; but although JSON::ize was a declared prerequisite, it's actually unused in REST::Neo4p.